+++ vite.config.ts
@@ -1,6 +1,7 @@
 import { resolve } from 'path'
 import { defineConfig } from 'vite'
 import { viteStaticCopy } from 'vite-plugin-static-copy'
+import removeConsole from 'vite-plugin-remove-console'
 import vue from '@vitejs/plugin-vue'
 import dts from "vite-plugin-dts"
 
@@ -24,7 +25,8 @@ export default defineConfig(({mode}) => ({
           dest: './'
         },
       ]
-    })
+    }),
+    removeConsole({ includes: ['log', 'info', 'warn', 'error', 'debug']})
   ],
   root: mode === 'development' ? 'playground' : '',
   resolve: {